Extract the variables in a graph
There is a bit of a hack in the admixture proportions that I need to fix in the interface to how graphs are constructed -- I check for parentheses to only get one of the edge probabilities but I allow the names of proportions to be different. I need to change that.