Interface to fitted data up again.
Most of the interface for accessing fitted data is functional again,
only predict I destroyed because it was actually pretty inconvenient to
make it work. Maybe contour_plot (which now works with more than 2
variables, user chooses 2) should count as well, so as to make it use
the class somehow? The algorithm is now faster as nelder mead uses a
fast lane and details are only calculated for the optimal solution.
Functions commented, lines cut shorter. Left taking the standard
deviation into account for next week.