Changed what is plotted...
The contour plot doesn’t re-estimate the edges any more but uses the
best fitting values for everything except the variables we vary.

Before it refitted the edges but not the admixture proportions. The
best might be to reestimate all variables in each point but that will
be slow — if we are not estimating admixture proportions it is more
consistent not to estimate the edges either.